# swagger/doc.yml
swagger: "2.0"
info:
  description: "A GinBro RESTful APIs"
  version: "1.0.0"
  title: "GinBro RESTful APIs Application"
host: "ynov-api.ew.r.appspot.com"
basePath: "/api/v1"
schemes:
- "http"
paths:
...
  /real-estate-propal:
    get:
      tags:
      - "real-estate-propal"
      summary: "get all real-estate-propal by pagination"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        default: "2"
        description: "limit returning object count"
        required: false
        type: "integer"

      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/RealEstatePropalPagination"
    post:
      tags:
      - "real-estate-propal"
      summary: "create real-estate-propal"
      description: "create real-estate-propal"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create real-estate-propal"
        required: true
        schema:
          $ref: "#/definitions/RealEstatePropal"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"

    patch:
      tags:
      - "real-estate-propal"
      summary: "update real-estate-propal"
      description: "update real-estate-propal"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create real-estate-propal"
        required: true
        schema:
          $ref: "#/definitions/RealEstatePropal"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /real-estate-propal/{Id}:
    get:
      tags:
      - "real-estate-propal"
      summary: "get a real-estate-propal by ID"
      description: "get a real-estate-propal by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of real-estate-propal to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/RealEstatePropal"
    delete:
      tags:
      - "real-estate-propal"
      summary: "Destroy a real-estate-propal by ID"
      description: "delete a real-estate-propal by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of real-estate-propal to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /real-estate-propals-buyer-indication:
    get:
      tags:
      - "real-estate-propals-buyer-indication"
      summary: "get all real-estate-propals-buyer-indication by pagination"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        default: "2"
        description: "limit returning object count"
        required: false
        type: "integer"

      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/RealEstatePropalsBuyerIndicationPagination"
    post:
      tags:
      - "real-estate-propals-buyer-indication"
      summary: "create real-estate-propals-buyer-indication"
      description: "create real-estate-propals-buyer-indication"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create real-estate-propals-buyer-indication"
        required: true
        schema:
          $ref: "#/definitions/RealEstatePropalsBuyerIndication"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"

    patch:
      tags:
      - "real-estate-propals-buyer-indication"
      summary: "update real-estate-propals-buyer-indication"
      description: "update real-estate-propals-buyer-indication"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create real-estate-propals-buyer-indication"
        required: true
        schema:
          $ref: "#/definitions/RealEstatePropalsBuyerIndication"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /real-estate-propals-buyer-indication/{Id}:
    get:
      tags:
      - "real-estate-propals-buyer-indication"
      summary: "get a real-estate-propals-buyer-indication by ID"
      description: "get a real-estate-propals-buyer-indication by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of real-estate-propals-buyer-indication to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/RealEstatePropalsBuyerIndication"
    delete:
      tags:
      - "real-estate-propals-buyer-indication"
      summary: "Destroy a real-estate-propals-buyer-indication by ID"
      description: "delete a real-estate-propals-buyer-indication by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of real-estate-propals-buyer-indication to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /real-estate-visit:
    get:
      tags:
      - "real-estate-visit"
      summary: "get all real-estate-visit by pagination"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        default: "2"
        description: "limit returning object count"
        required: false
        type: "integer"

      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/RealEstateVisitPagination"
    post:
      tags:
      - "real-estate-visit"
      summary: "create real-estate-visit"
      description: "create real-estate-visit"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create real-estate-visit"
        required: true
        schema:
          $ref: "#/definitions/RealEstateVisit"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"

    patch:
      tags:
      - "real-estate-visit"
      summary: "update real-estate-visit"
      description: "update real-estate-visit"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create real-estate-visit"
        required: true
        schema:
          $ref: "#/definitions/RealEstateVisit"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /real-estate-visit/{Id}:
    get:
      tags:
      - "real-estate-visit"
      summary: "get a real-estate-visit by ID"
      description: "get a real-estate-visit by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of real-estate-visit to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/RealEstateVisit"
    delete:
      tags:
      - "real-estate-visit"
      summary: "Destroy a real-estate-visit by ID"
      description: "delete a real-estate-visit by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of real-estate-visit to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
...
  /user:
    get:
      tags:
      - "user"
      summary: "get all user by pagination"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        default: "2"
        description: "limit returning object count"
        required: false
        type: "integer"

      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/UserPagination"
    post:
      tags:
      - "user"
      summary: "create user"
      description: "create user"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create user"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"

    patch:
      tags:
      - "user"
      summary: "update user"
      description: "update user"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create user"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /user/{Id}:
    get:
      tags:
      - "user"
      summary: "get a user by ID"
      description: "get a user by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of user to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/User"
    delete:
      tags:
      - "user"
      summary: "Destroy a user by ID"
      description: "delete a user by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of user to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
definitions:
...
  RealEstatePropal:
    type: "object"
    properties:
    
      id:
        type: "integer"
        description: "Id de la proposition"
        format: "int64"
      
      id_booker:
        type: "integer"
        description: "Id du booker"
        format: "int64"
      
      id_buyer:
        type: "integer"
        description: "Id du visitor"
        format: "int64"
      
      id_real_estate:
        type: "integer"
        description: "Id du bien"
        format: "int64"
      
      id_vendor:
        type: "integer"
        description: "Id du propriétaire"
        format: "int64"
      
      price:
        type: "number"
        description: "Montant de la proposition"
        format: "float"
      
      buyer_first_name:
        type: "string"
        description: "Prénom de l'acheteur"
        format: "string"
      
      buyer_last_name:
        type: "string"
        description: "Nom de l'acheteur"
        format: "string"
      
      buyer_address:
        type: "string"
        description: "Adresse de l'acheteur"
        format: "string"
      
      buyer_zip_code:
        type: "string"
        description: "Code postal de l'acheteur"
        format: "string"
      
      buyer_city:
        type: "string"
        description: "Ville de l'acheteur"
        format: "string"
      
      expiration_date:
        type: "string"
        description: "Date d'expiration de l'offre"
        format: "date-time"
      
      is_accepted:
        type: "integer"
        description: "Est accepté?"
        format: "int64"
  RealEstatePropalPagination:
    type: "object"
    properties:
      code:
        type: "integer"
        description: "json repose code"
        format: "int32"
      total:
        type: "integer"
        description: "total numbers"
        format: "int32"
      offset:
        type: "integer"
        description: "offset"
        format: "int32"
      limit:
        type: "integer"
        description: "limit"
        format: "int32"
      list:
        type: "array"
        items:
          $ref: "#/definitions/RealEstatePropal"
  RealEstatePropalsBuyerIndication:
    type: "object"
    properties:
    
      id:
        type: "integer"
        description: "Id de l'indication"
        format: "int64"
      
      id_propal:
        type: "integer"
        description: "Id de la proposition"
        format: "int64"
      
      id_buyer:
        type: "integer"
        description: "Id de l'acheteur"
        format: "int64"
      
      id_vendor:
        type: "integer"
        description: "Id du vendeur"
        format: "int64"
      
      has_total_in_cash:
        type: "integer"
        description: "a le total en cash ?"
        format: "int64"
      
      has_financial_support:
        type: "integer"
        description: "a un apport financier ?"
        format: "int64"
      
      has_big_financial_support:
        type: "integer"
        description: "a un apport financier important ? (>30% du bien)"
        format: "int64"
      
      has_very_big_financial_support:
        type: "integer"
        description: "a un apport financier très important ? (> 70% du bien)"
        format: "int64"
      
      want_a_long_sell:
        type: "integer"
        description: "veut une vente longue ? (+ de 6 mois)"
        format: "int64"
  RealEstatePropalsBuyerIndicationPagination:
    type: "object"
    properties:
      code:
        type: "integer"
        description: "json repose code"
        format: "int32"
      total:
        type: "integer"
        description: "total numbers"
        format: "int32"
      offset:
        type: "integer"
        description: "offset"
        format: "int32"
      limit:
        type: "integer"
        description: "limit"
        format: "int32"
      list:
        type: "array"
        items:
          $ref: "#/definitions/RealEstatePropalsBuyerIndication"
  RealEstateVisit:
    type: "object"
    properties:
    
      id:
        type: "integer"
        description: "Id de la visite"
        format: "int64"
      
      id_real_estate:
        type: "integer"
        description: "Id du bien"
        format: "int64"
      
      id_booker:
        type: "integer"
        description: "Id du booker"
        format: "int64"
      
      id_visitor:
        type: "integer"
        description: "Id du visiteur"
        format: "int64"
      
      start_date:
        type: "string"
        description: "Date de début (YYYY-mm-dd)"
        format: "date-time"
      
      end_date:
        type: "string"
        description: "Date de fin (YYYY-mm-dd)"
        format: "date-time"
      
      start_time:
        type: "string"
        description: "Heure de début (hh:mm)"
        format: "date-time"
      
      end_time:
        type: "string"
        description: "Heure de fin (hh:mm)"
        format: "date-time"
      
      booker_is_ready:
        type: "integer"
        description: "Le booker est prêt ?"
        format: "int64"
      
      visitor_is_ready:
        type: "integer"
        description: "Les visiteurs sont prêts ?"
        format: "int64"
  RealEstateVisitPagination:
    type: "object"
    properties:
      code:
        type: "integer"
        description: "json repose code"
        format: "int32"
      total:
        type: "integer"
        description: "total numbers"
        format: "int32"
      offset:
        type: "integer"
        description: "offset"
        format: "int32"
      limit:
        type: "integer"
        description: "limit"
        format: "int32"
      list:
        type: "array"
        items:
          $ref: "#/definitions/RealEstateVisit"
...
  User:
    type: "object"
    properties:
    
      id:
        type: "integer"
        description: "Id de l'utilisateur"
        format: "int64"
      
      email:
        type: "string"
        description: "Email de l'utilisateur qui sert aussi de login"
        format: "string"
      
      password:
        type: "string"
        description: "Mot de passe de l'utilisateur"
        format: "string"
      
      avatar:
        type: "string"
        description: "Url de l'avatar de l'utilisateur"
        format: "string"
      
      first_name:
        type: "string"
        description: "Prénom de l'utilisateur"
        format: "string"
      
      last_name:
        type: "string"
        description: "Nom de l'utilisateur"
        format: "string"
      
      address:
        type: "string"
        description: "Adresse"
        format: "string"
      
      zip_code:
        type: "string"
        description: "Code postal"
        format: "string"
      
      city:
        type: "string"
        description: "Ville"
        format: "string"
      
      latitude:
        type: "string"
        description: "Latitude"
        format: "string"
      
      longitude:
        type: "string"
        description: "Longitude"
        format: "string"
  UserPagination:
    type: "object"
    properties:
      code:
        type: "integer"
        description: "json repose code"
        format: "int32"
      total:
        type: "integer"
        description: "total numbers"
        format: "int32"
      offset:
        type: "integer"
        description: "offset"
        format: "int32"
      limit:
        type: "integer"
        description: "limit"
        format: "int32"
      list:
        type: "array"
        items:
          $ref: "#/definitions/User"
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      msg:
        type: "string"
...